index.phpと.htaccessの書き換え＆移動に際し条件分岐を追加

+++ docker-entrypoint.sh
@@ -281,8 +281,10 @@ EOPHP
 	done
 fi
 
-sed -i -e 's/\/wp-blog-header\.php/\/wordpress\/wp-blog-header.php/g' /var/www/html/wordpress/index.php
-mv /var/www/html/wordpress/.htaccess /var/www/html/.htaccess
-mv /var/www/html/wordpress/index.php /var/www/html/index.php
+if [ -e /var/www/html/wordpress/index.php ] && [ -e /var/www/html/wordpress/.htaccess ] [ ! -e /var/www/html/index.php ] && [ ! -e /var/www/html/.htaccess ]; then
+  sed -i -e 's/\/wp-blog-header\.php/\/wordpress\/wp-blog-header.php/g' /var/www/html/wordpress/index.php
+  mv /var/www/html/wordpress/.htaccess /var/www/html/.htaccess
+  mv /var/www/html/wordpress/index.php /var/www/html/index.php
+fi
 
 exec "$@"
